+++ docs/CHANGELOG.md
@@ -20,6 +20,7 @@ Changelog
 - Enh #38: Save menu collapse state in local storage
 - Fix #50: Right arrow navigation not working in some cases
 - Enh #102: Source Button Position
+- Enh #104: Source Mode, Text Area Style
 
 
 1.1.3 (Unreleased)